LRCI-3781 Move Job Information into its own component
